move fixtures to app-review
