+++ packages/@coorpacademy-app-review/src/actions/api/test/fetch-correction.test.ts
@@ -2,7 +2,6 @@ import test from 'ava';
 import type {Services} from '@coorpacademy/review-services';
 import {
   getChoicesCorrection,
-  postAnswerResponses,
   services as mockedServices
 } from '@coorpacademy/review-services-mocks';
 import type {StoreState} from '../../../reducers';
@@ -15,6 +14,7 @@ import {
 } from '../fetch-correction';
 import {freeTextSlide} from '../../../views/slides/test/fixtures/free-text';
 import {qcmSlide} from '../../../views/slides/test/fixtures/qcm';
+import {postAnswerResponses} from '../../../test/fixtures';
 
 const progression = postAnswerResponses[freeTextSlide.universalRef];
 const skillRef = '_skill-ref';
