+++ packages/@coorpacademy-app-review/src/views/slides/test/index.test.ts
@@ -9,14 +9,13 @@ import {CongratsCardProps} from '@coorpacademy/components/es/molecule/review-car
 
 import {
   postProgressionResponse as createdProgression,
-  postAnswerResponses,
   progressionSlideWithPendingSlide,
-  getChoicesCorrection,
-  incorrectFreeTextPostAnswerResponse
+  getChoicesCorrection
 } from '@coorpacademy/review-services-mocks';
 import {mapStateToSlidesProps} from '..';
 import type {StoreState} from '../../../reducers';
 import {translate} from '../../../../sandbox/translation.mock';
+import {incorrectFreeTextPostAnswerResponse, postAnswerResponses} from '../../../test/fixtures';
 import {skin} from './fixtures/skin';
 import {sliderSlide} from './fixtures/slider';
 import {templateSlide} from './fixtures/template';
