+++ packages/@coorpacademy-app-review/src/views/slides/test/slide.next-slide.on-click.test.ts
@@ -2,17 +2,13 @@ import test from 'ava';
 import identity from 'lodash/fp/identity';
 import {ReviewCorrectionPopinProps} from '@coorpacademy/components/es/molecule/review-correction-popin/prop-types';
 
-import {
-  postAnswerResponses,
-  getChoicesCorrection,
-  incorrectFreeTextPostAnswerResponse,
-  services
-} from '@coorpacademy/review-services-mocks';
+import {getChoicesCorrection, services} from '@coorpacademy/review-services-mocks';
 import type {StoreState} from '../../../reducers';
 import {mapStateToSlidesProps} from '..';
 import {createTestStore} from '../../../actions/test/create-test-store';
 import {NEXT_SLIDE} from '../../../actions/ui/next-slide';
 import {translate} from '../../../../sandbox/translation.mock';
+import {incorrectFreeTextPostAnswerResponse, postAnswerResponses} from '../../../test/fixtures';
 import {sliderSlide} from './fixtures/slider';
 import {skin} from './fixtures/skin';
 import {freeTextSlide} from './fixtures/free-text';
